requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
pymongo==3.12.3
pytz==2023.3
sqlparse==0.2.4
tzdata==2023.3
